+++ lib/AppInfo/Application.php
@@ -32,6 +32,7 @@ use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
 use OCP\AppFramework\Bootstrap\IRegistrationContext;
+use OCP\User\Events\UserChangedEvent;
 
 final class Application extends App implements IBootstrap {
 	public const APP_ID = 'twofactor_email';
@@ -54,6 +55,7 @@ final class Application extends App implements IBootstrap {
 		$context->registerEventListener(StateChanged::class, StateChangeActivity::class);
 		$context->registerEventListener(StateChanged::class, StateChangeRegistryUpdater::class);
 		$context->registerEventListener(UserUpdatedEvent::class, EMailDeleted::class);
+		$context->registerEventListener(UserChangedEvent::class, EMailDeleted::class);
 	}
 
 	public function boot(IBootContext $context): void {