+++ tests/Unit/Listener/StateChangeActivityTest.php
@@ -10,10 +10,12 @@ declare(strict_types=1);
 namespace OCA\TwoFactorEMail\Test\Unit\Listener;
 
 use ChristophWurst\Nextcloud\Testing\TestCase;
+use OCA\TwoFactorEMail\Event\StateChangeActor;
 use OCA\TwoFactorEMail\Event\StateChanged;
 use OCA\TwoFactorEMail\Listener\StateChangeActivity;
 use OCP\Activity\IEvent;
 use OCP\Activity\IManager;
+use OCP\EventDispatcher\Event;
 use OCP\IUser;
 use PHPUnit\Framework\MockObject\Exception;
 use PHPUnit\Framework\MockObject\MockObject;
@@ -57,6 +59,34 @@ class StateChangeActivityTest extends TestCase {
 			->with($activityEvent);
 
 		$this->listener->handle($event);
+	}
+
+	/**
+	 * @throws Exception
+	 */
+	public function testSystemDisableCreatesNoEmailActivity(): void {
+		$user = $this->createMock(IUser::class);
+		$user->method('getUID')->willReturn('user234');
+		$event = new StateChanged($user, false, StateChangeActor::SYSTEM);
+		$activityEvent = $this->createMock(IEvent::class);
+		$activityEvent->method('setApp')->willReturnSelf();
+		$activityEvent->method('setType')->willReturnSelf();
+		$activityEvent->method('setAuthor')->willReturnSelf();
+		$activityEvent->method('setAffectedUser')->willReturnSelf();
+		$activityEvent->expects($this->once())
+			->method('setSubject')
+			->with('twofactor_email_disabled_no_email')
+			->willReturnSelf();
+		$this->activityManager->method('generateEvent')->willReturn($activityEvent);
+		$this->activityManager->expects($this->once())->method('publish');
+
+		$this->listener->handle($event);
+	}
+
+	public function testIgnoresForeignEvents(): void {
+		$this->activityManager->expects($this->never())->method('generateEvent');
+
+		$this->listener->handle(new Event());
 	}
 
 	/**